feat: Endles mode Added

+++ src/model/wall/Wall.java
@@ -6,6 +6,7 @@ import javafx.scene.Node;
 import javafx.scene.image.Image;
 import model.Entity;
 import model.GameObject;
+import model.player.Player;
 
 import java.util.ArrayList;
 
@@ -18,7 +19,7 @@ public class Wall extends GameObject {
 
     public static boolean canMoveUp(Entity entity, ArrayList<Wall> wallArrayList) {
 
-        return canMove(entity, wallArrayList, false, -MapLoader.BLOCK_SIZE);
+        return canMove(entity, wallArrayList, false, -(int)MapLoader.BLOCK_SIZE);
     }
 
     public static boolean canMoveDown(Entity entity, ArrayList<Wall> wallArrayList) {
@@ -26,12 +27,13 @@ public class Wall extends GameObject {
     }
 
     public static boolean canMoveLeft(Entity entity, ArrayList<Wall> wallArrayList) {
-        return canMove(entity, wallArrayList, true, -MapLoader.BLOCK_SIZE);
+        return canMove(entity, wallArrayList, true, -(int)MapLoader.BLOCK_SIZE);
     }
 
     public static boolean canMoveRight(Entity entity, ArrayList<Wall> wallArrayList) {
         return canMove(entity, wallArrayList, true, (int) entity.getImage().getHeight());
     }
+
 
     public static boolean canMove(GameObject gameObject, ArrayList<Wall> wallArrayList, boolean horizontal, double offset) {
         if (wallArrayList == null)
